+++ testapps/GH918/src/main/java/gh918/AppEntry.java
@@ -1,0 +1,15 @@
+package gh918;
+
+import act.Act;
+import org.osgl.mvc.annotation.GetAction;
+
+public class AppEntry {
+
+    @GetAction
+    public void test() {
+    }
+
+    public static void main(String[] args) throws Exception {
+        Act.start();
+    }
+}